Avoid using the identifier "R10", for the benefit of MinGW.
The symptom reported in #30 is:

    g++ -std=c++11 -mwindows hyper.cpp hyper.res -o hyper.exe \
        -lSDL -lSDL_mixer -lopengl32 -lSDL_ttf -lSDL_gfx -lglew32 \
        -O3 -DWINDOWS -D_A_VOLID=8

    In file included from init.cpp:337:0,
                     from hyper.cpp:34:
    classes.cpp:1837:22: error: expected ')' before '?' token
     #define R10 (inv::on ? 50 : 10)
                          ^

I believe what's going on here is that this is a particularly
unfriendly compiler error message. I believe the actual error is
somewhere else, unmentioned in the message, such as MinGW-w64's
headers "crt/setjmp.h" or "winnt.h", which define struct fields
named `R10`. Our macro is getting defined *first*, and then
expanded inside those system headers when they're included
(somehow, whenever that is), which produces a syntax error of
the form

    DWORD64 R9;
    DWORD64 (inv::on ? 50 : 10);
    DWORD64 R11;

The compiler is just not showing us this part of the story.
So, trivial fix: just rename all the `R*` macros to `R_*`.
No intended functional change.