MSVC warns about bool/int conversion.
This patch does, technically, fix the following MSVC diagnostics.
But I believe that it should *not* be committed exactly as-is;
I suspect that some of these diagnostics — especially the ones in
`mapeditor.cpp` — might indicate possible bugs or at least tech debt.
Committing this patch as-is would just make the bug/debt harder to find
in the future.

There is also a trivial workaround for MSVC: I can just pass `/wd4805`,
the same way I pass `-Wno-unused-result` on GCC.

    heptagon.cpp(201): warning C4805: '^': unsafe mix of type 'const bool' and type 'int' in operation
    pattern2.cpp(627): warning C4805: '^=': unsafe mix of type 'int' and type 'bool' in operation
    pattern2.cpp(669): warning C4805: '^=': unsafe mix of type 'int' and type 'bool' in operation
    mapeditor.cpp(1325): warning C4805: '!=': unsafe mix of type 'int' and type 'bool' in operation
    mapeditor.cpp(1770): warning C4805: '==': unsafe mix of type 'bool' and type 'int' in operation
    hypgraph.cpp(583): warning C4805: '^': unsafe mix of type 'const bool' and type 'int' in operation
    hyper.h(2025): warning C4805: '!=': unsafe mix of type 'bool' and type 'int' in operation
    graph.cpp(3464): note: see reference to function template instantiation 'V hr::callhandlers<bool(hr::cell *,const hr::transmatrix &),int,hr::cell*,hr::transmatrix>(V,hr::hookset<bool (hr::cell *,const hr::transmatrix &)> *,hr::cell *&,hr::transmatrix &)' being compiled with [V=int]