Fix a couple of issues in MSVC related to `min` and `max`.
MSVC apparently thinks that the result of arithmetic promotions
on an `unsigned:4` bitfield is `unsigned`, not `int`; which then
causes it to fail to deduce whether the `T` in `min<T>` should be
`unsigned` (the type of the LHS) or `int` (the type of the RHS).
Clang and GCC agree that the result of arithmetic promotions on
an `unsigned:4` bitfield should be `int`, so they don't see any
ambiguity here.