Add `colorpartproxy` so that we can write setpart(c,i)=v instead of setpart(c,i,v)
No intended functional change.
Notice that this is very conservative and deliberately does not break
things like

    const auto& pix = part(c, i);

That is, `part` still returns a prvalue `unsigned char`, and if you want to
modify it, you use `setpart`. All this patch does is let you use the `=` sign
at the appropriate place in the statement, for readability.