langen: Handle repeated strings better.
If a translation file contains two copies of the same English string,
with the same translation, then we don't care at all; just throw out
the second copy.

If a translation file contains two copies of the same English string,
with DIFFERENT translations, then (according to how the game does
translation) one of those translations must be inappropriate. Give
a warning in that case; and again, throw out the second copy because
we are forced to pick ONE translation for use within the game.

Also fix some minor typos I noticed in the strings.