Add minetexts up to 14
The truncated heptagonal and truncated triheptagonal tilings both appear in
the list of Archimedean tilings, and both contain tetradecagons. It seems that
this is the most sides of any polygon in any of the game's non-custom tilings,
so it seems like a good choice for how high minetexts should support.